# Add aggreate key dependency optimizer