--- a/src/planner/operator/logical_aggregate.cpp
+++ b/src/planner/operator/logical_aggregate.cpp
@@ -11,28 +11,18 @@ using namespace factorization;
 void LogicalAggregate::computeFactorizedSchema() {
     createEmptySchema();
     auto groupPos = schema->createGroup();
-    for (auto& expression : expressionsToGroupBy) {
-        schema->insertToGroupAndScope(expression, groupPos);
-    }
-    for (auto& expression : expressionsToAggregate) {
-        schema->insertToGroupAndScope(expression, groupPos);
-    }
+    insertAllExpressionsToGroupAndScope(groupPos);
 }
 
 void LogicalAggregate::computeFlatSchema() {
     createEmptySchema();
     schema->createGroup();
-    for (auto& expression : expressionsToGroupBy) {
-        schema->insertToGroupAndScope(expression, 0);
-    }
-    for (auto& expression : expressionsToAggregate) {
-        schema->insertToGroupAndScope(expression, 0);
-    }
+    insertAllExpressionsToGroupAndScope(0 /* groupPos */);
 }
 
 f_group_pos_set LogicalAggregate::getGroupsPosToFlattenForGroupBy() {
     f_group_pos_set dependentGroupsPos;
-    for (auto& expression : expressionsToGroupBy) {
+    for (auto& expression : getAllKeyExpressions()) {
         for (auto groupPos : children[0]->getSchema()->getDependentGroupsPos(expression)) {
             dependentGroupsPos.insert(groupPos);
         }
@@ -48,7 +38,7 @@ f_group_pos_set LogicalAggregate::getGroupsPosToFlattenForGroupBy() {
 f_group_pos_set LogicalAggregate::getGroupsPosToFlattenForAggregate() {
     if (hasDistinctAggregate()) {
         f_group_pos_set dependentGroupsPos;
-        for (auto& expression : expressionsToAggregate) {
+        for (auto& expression : aggregateExpressions) {
             for (auto groupPos : children[0]->getSchema()->getDependentGroupsPos(expression)) {
                 dependentGroupsPos.insert(groupPos);
             }
@@ -60,11 +50,14 @@ f_group_pos_set LogicalAggregate::getGroupsPosToFlattenForAggregate() {
 
 std::string LogicalAggregate::getExpressionsForPrinting() const {
     std::string result = "Group By [";
-    for (auto& expression : expressionsToGroupBy) {
+    for (auto& expression : keyExpressions) {
+        result += expression->toString() + ", ";
+    }
+    for (auto& expression : dependentKeyExpressions) {
         result += expression->toString() + ", ";
     }
     result += "], Aggregate [";
-    for (auto& expression : expressionsToAggregate) {
+    for (auto& expression : aggregateExpressions) {
         result += expression->toString() + ", ";
     }
     result += "]";
@@ -72,13 +65,25 @@ std::string LogicalAggregate::getExpressionsForPrinting() const {
 }
 
 bool LogicalAggregate::hasDistinctAggregate() {
-    for (auto& expressionToAggregate : expressionsToAggregate) {
-        auto& functionExpression = (binder::AggregateFunctionExpression&)*expressionToAggregate;
+    for (auto& expression : aggregateExpressions) {
+        auto& functionExpression = (binder::AggregateFunctionExpression&)*expression;
         if (functionExpression.isDistinct()) {
             return true;
         }
     }
     return false;
+}
+
+void LogicalAggregate::insertAllExpressionsToGroupAndScope(f_group_pos groupPos) {
+    for (auto& expression : keyExpressions) {
+        schema->insertToGroupAndScope(expression, groupPos);
+    }
+    for (auto& expression : dependentKeyExpressions) {
+        schema->insertToGroupAndScope(expression, groupPos);
+    }
+    for (auto& expression : aggregateExpressions) {
+        schema->insertToGroupAndScope(expression, groupPos);
+    }
 }
 
 } // namespace planner