… src/optimizer/CMakeLists.txt
… src/processor/mapper/map_aggregate.cpp
… src/processor/mapper/map_distinct.cpp
… src/processor/operator/aggregate/aggregate_hash_table.cpp
… src/processor/operator/aggregate/hash_aggregate.cpp
… test/test_files/tinysnb/agg/multi_query_part.test