This PR adds AggregateKeyDependencyOptimizer which analyzes user input group by keys' dependency and split into `keys` and `dependentKeys`. A `dependentKey` has functional dependency on a `key`. `keys` will be hashed during hash aggregate while `dependentKeys` will be directly materialized into f-table. 

Consider the following example
```
RETURN a.ID, a.age, COUNT(*)
```

Based on user input, both `a.ID` and `a.age` are stated as hash keys. However, since `a.ID` is the primary key, `a.age` has functional dependency on `a.ID` so we only need to hash `a.ID` in hash aggregate.

This optimizer is more important given Cypher grammar because the grammar itself doesn't differentiate `SELECT` and `GROUP BY` which means user cannot group by a subset of selection list. Another case is Cypher allows group by `Node` or `Rel` which by definition is to group by all properties of `Node` or `Rel`. A more reasonable solution is to only group by their internal IDs and treat properties as dependentKeys.

## Performance benchmark
Dataset: LDBC100
Machine: M1 Pro, 16GB memory, maximum 4 threads
Query: `MATCH (a:Person)-[:knows]->(b:Person) RETURN keys, COUNT(*) LIMIT 1;` (Note that Aggregate is a Sink operator that is not affected by LIMIT, LIMIT 1 is to reduce print size).

| Keys | Optimizer ON    | Optimizer OFF |
| ----- | -------- | ------- |
| ID | 207 ms  | 198 ms   |
| ID, firstName| 204 ms  | 292 ms   |
| ID, firstName, lastName | 214 ms  | 356 ms   |